+++ webapp/vite.config.ts
@@ -4,6 +4,7 @@ import react from '@vitejs/plugin-react';
 export default defineConfig({
   plugins: [react()],
   server: {
+    allowedHosts: ['goliath'],
     port: 11017,
     proxy: {
       '/api': 'http://127.0.0.1:11016',